Address Sleepy Softlock review: Softlocks folder, trim comments, logging.
Move the patch under Patches/Softlocks, extract the Dreamer scene check, and wrap the repair in try/catch with debug/info logs. Harmony still patches YieldRestart globally; the scene/name guards keep work chapter-scoped.
